move di related stuff to separate directory
