+++ app/components/App.tsx
@@ -2,11 +2,11 @@ import * as React from "react";
 import { connect } from "react-redux";
 import { ToastContainer } from "react-toastify";
 
+import { symbols } from "../di/symbols";
 import { flows } from "../modules/flows";
 import { Web3Manager } from "../modules/web3/Web3Manager";
 import { injectableFn } from "../redux-injectify";
 import { AppDispatch } from "../store";
-import { symbols } from "../symbols";
 import { AppRouter } from "./AppRouter";
 import { LoadingIndicator } from "./shared/LoadingIndicator";
