+++ app/modules/kyc/sagas.ts
@@ -4,8 +4,8 @@ import { actions, TAction } from "../actions";
 
 import { getDependency, neuTake } from "../sagas";
 
+import { symbols } from "../../di/symbols";
 import { IApiKycService } from "../../lib";
-import { symbols } from "../../symbols";
 
 function* submitCompanyForm(): Iterator<any> {
   while (true) {
