+++ app/modules/networking/SignatureAuthApi.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from "inversify";
 import * as Yup from "yup";
 
-import { symbols } from "../../symbols";
+import { symbols } from "../../di/symbols";
 import { EthereumAddressWithChecksum } from "../../types";
 import { SignerType } from "../web3/PersonalWeb3";
 import { IHttpClient, IHttpResponse } from "./IHttpClient";
