+++ app/modules/userAgent/flows.ts
@@ -1,8 +1,8 @@
 import * as detectBrowser from "detect-browser";
 
+import { symbols } from "../../di/symbols";
 import { injectableFn } from "../../redux-injectify";
 import { AppDispatch } from "../../store";
-import { symbols } from "../../symbols";
 import { actions } from "../actions";
 
 export const userAgentFlows = {
